requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==1.0.4
langchain-core==1.0.4
langchain-google-genai==3.0.3
vnstock==3.3.0
pandas==2.2.3
numpy==2.3.4
pandas-ta-classic==0.3.37
python-dotenv==1.0.1
pydantic==2.10.6
tabulate==0.9.0
langgraph==1.0.3